+++ statistics.c
@@ -64,19 +64,19 @@ void statistics_main(GtkBuilder* ui)
     g_string_append_printf(str, "Today's date: %s, %02d.%02d.%d\n\n", day_to_string(currentdweek), currentday, currentmonth, currentyear);
 
     struct db_entry* entries = db_get_all();
-    long long past_sum = 0;
-    long long future_sum = 0;
+    double past_sum = 0;
+    double future_sum = 0;
     for (int i = 0; entries[i].name[0] != '\0'; i++) {
         if (entries[i].month == currentmonth && entries[i].year == currentyear && entries[i].day <= currentday)
             past_sum += entries[i].price;
         if (entries[i].month == currentmonth && entries[i].year == currentyear && entries[i].day > currentday)
             future_sum += entries[i].price;
     }
-    g_string_append_printf(str, "Spent until now this month: %lld\n", past_sum);
+    g_string_append_printf(str, "Spent until now this month: %lf\n", past_sum);
 
-    g_string_append_printf(str, "Sum of planned future spendings this month: %lld\n", future_sum);
+    g_string_append_printf(str, "Sum of planned future spendings this month: %lf\n", future_sum);
 
-    g_string_append_printf(str, "Total sum of money planned to be spent this month : %lld\n", future_sum + past_sum);
+    g_string_append_printf(str, "Total sum of money planned to be spent this month : %lf\n", future_sum + past_sum);
 
     gtk_label_set_text(label, str->str);
 }
